pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "whl-cal"
version = "0.1.0"
description = "LiDAR extrinsic calibration tool"
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
authors = [
  { name = "spidermaniron", email = "1195487354@qq.com" }
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]
dependencies = [
  "click",
  "numpy",
  "open3d",
  "protobuf>=5.29,<6",
  "pycyber",
  "scipy",
  "pyyaml"
]

[project.urls]
Homepage = "https://github.com/wheelos-tools/whl-cal"
BugTracker = "https://github.com/wheelos-tools/whl-cal/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
lidar2lidar = ["conf/*.yaml", "conf/*.md"]

[project.scripts]
lidar2lidar-calibrate = "lidar2lidar.cli:main"
lidar2lidartool = "lidar2lidar.cli:main"
lidar2lidar-auto = "lidar2lidar.auto_calib:main"
lidar2lidar-scan2map = "lidar2lidar.scan2map_calib:main"
lidar2lidar-scan2map-dataset = "lidar2lidar.scan2map_dataset:main"
lidar2lidar-temporal = "lidar2lidar.temporal_calib:main"
lidar2imu-calibrate = "lidar2imu.cli:main"
lidar2imu-convert-record = "lidar2imu.record_converter:main"
lidar2imu-tune-record = "tools.lidar2imu.tune_record:main"
lidar2lidar-extract = "tools.lidar2lidar.extract_pcd:main"
lidar2lidar-topics = "tools.lidar2lidar.list_topics:main"
lidar2lidar-merge = "tools.lidar2lidar.merge_pcd:main"
